+++ src/main/java/com/pfe/pfe/dto/RegisterEmployeRequest.java
@@ -11,7 +11,7 @@ public class RegisterEmployeRequest {
 
     private String nomUtil ;
     private String prenomUtil ;
-    private int numTel ;
+    private long numTel;
     private String emailUtil ;
     private String motDePasse ;
